将管理员列表的类型更改为long

+++ src/main/java/cn/pac/zeros/Config.java
@@ -26,7 +26,7 @@ public class Config {
     public class ConfigObject {
         public Group Group;
         public class Group {
-            public int[] Admin;// 管理员群组
+            public long[] Admin;// 管理员群组
             public int[] Common;// 交流群组
         }
         public Server Server;
